Commonise code for different release branches
Refactor release versions script to reduce the number of places
in the code that need to check if a given release branch exists
or not.

Commonise the code that fetches the release branch details, and
the code that generates the view model data to be used in the HTML
template.