Handle case where some release branches are missing
This fixes a bug where the release version script would fail because
a new repo had been deployed to integration but not staging or prod.

Show the release version as "Unknown" in the table, and highlight it
in red.

The None handling is quite messy, so the next commit will refactor the
code to make the error handling simpler.