Use the new subnet for the bastion.
The old subnet was created by the EFS module in terraform. Now that the
backend checks EFS is being removed, the subnet is being moved to a new
module and it has a new name that doesn't have -efs- in it.

This needs updating to use the new name.

I've also removed the link to the backend checks bastion and the option
to connect to it on the GitHub actions workflow.